Add `Internal` module
